Exceptions reworked + yml config

+++ common/src/main/java/com.luchkovskiy/service/implementations/AggregationServiceImpl.java
@@ -3,8 +3,11 @@ package com.luchkovskiy.service.implementations;
 import com.luchkovskiy.models.User;
 import com.luchkovskiy.repository.UserRepository;
 import com.luchkovskiy.service.AggregationService;
+import com.luchkovskiy.service.exceptions.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
 
 
 @Service
@@ -16,9 +19,9 @@ public class AggregationServiceImpl implements AggregationService {
     @Override
     public boolean deleteInactiveUser(Long userId) {
         if (!userRepository.existsById(userId))
-            throw new RuntimeException();
-        User user = userRepository.findById(userId).orElseThrow(RuntimeException::new);
-        if (!user.getActive()) {
+            throw new EntityNotFoundException("User not found!");
+        User user = userRepository.findById(userId).get();
+        if (!user.getActive() && user.getChanged().isBefore(LocalDateTime.now().minusMonths(1))) {
             userRepository.delete(user);
             return true;
         }
